Use macro instead of runtime AST manipulation.
