--- syntax/src/main/scala/org/enso/syntax/Main.scala
@@ -1,71 +1,0 @@
-package org.enso.syntax
-import scala.meta._
-import scala.annotation._
-
-object Main extends App {
-
-  var indent = 0
-
-  def pprint[T](t: T) {
-    if (t == null) {
-      println(t)
-      return
-    }
-    val s = t.toString()
-    print("  " * indent)
-    val (l, r2) = s.span(x => x != '(' && x != ')')
-    print(l)
-    if (r2 == "") {
-      println
-      return
-    }
-
-    val (m, r) = r2.splitAt(1)
-
-    if (m == "(") {
-      indent += 1
-      println(m)
-      pprint(r)
-    } else if (m == ")") {
-      indent -= 1
-      println(m)
-      pprint(r)
-    }
-
-  }
-
-////  val str = "a (b"
-//  val str = "a\n b\n a" // .stripMargin
-//  println(str)
-//  val reader = new StringReader(str)
-//  val ss     = new Lexer(reader)
-//  val toks   = ss.lexAll()
-//  var ttt    = 10
-//  pprint(toks.toString)
-//
-//  val sparser = new SParser(new StringReader(str))
-//
-//  val bparser = new BParser(new StringReader(str))
-//  val parser  = new ppp.Parser(new StringReader(str))
-//
-//  pprint(bparser.parse.toString())
-//  pprint(parser.parse.toString())
-//  pprint("!")
-//  println(sparser.strInput)
-//  pprint(sparser.parse.toString)
-
-  //////////////////////////////////////////////////////////////
-
-  val p1   = new Parser
-  val code = p1.specialize()
-  val p2   = p1.debugRun("'foo'")
-  pprint(p2)
-  p2 match {
-    case Flexer.Success(v, _) =>
-      println(v.span)
-  }
-  println("CODE LEN:", code.length) //136500
-  //  println(tree1)
-//  println(p2.result.show())
-
-}
