--- syntax/src/main/scala/org/enso/syntax/package.scala
@@ -1,7 +1,0 @@
-package org.enso
-
-import org.enso.syntax.AST
-
-package object syntax {
-  type AST = AST.AST
-}
